Fix a couple of synthetic properties
Currently, the drop downs to choose the kind of application manifest (custom, none, or default) and the kind of package license expression (none, expression, or file) are not working properly: when the user chooses picks a different option, the UI resets to the default value almost immediately.

The issue is that these are synthetic properties--they don't have any direct representation in the project, and instead compute their values based on the values of _other_ properties. If those properties don't yet have values, then the next time we retrieve the value of the synthetic property we will get its default. That's what's happening here: the user changes the property, the UI sends that property change and requests an update, it gets the default value back, and resets the UI.

The fix here is to store the synthetic property value in the `ITemporaryPropertyStorage` when setting it. The next time the property is retrieved we first try to compute it from other values. If that fails we try to get it from the `ITemporaryPropertyStorage`. If it hasn't been stored we return the default.