Add a provider for launch profile "types"
When using launch profiles for debugging, the most important property of a profile is the _command name_. This controls which component gets to take the active launch profile and turn it into debugger settings. As such, the command name effectively determines which properties in the profile are relevant to debugging, and which ones should be surfaced to the user in the UI and how.

We already capture most of this information in `Rule` files where the PageTemplate is "commandNameBasedDebugger" (aside: for clarity, we should consider changing this to "launchProfilePage" or similar).

In this commit we find each of these `Rule` files in the current project and expose them through the Project Query API as `LaunchProfileType` entities. These entities include the command name (contained in the `Rule` metadata), a display name (which comes from the `DisplayName` attribute of the `Rule`), a help URL, and an `ImageMoniker` (again, the data for these is included in the `Rule` metadata).

The image moniker allows us to surface different icons for different kinds of launch profiles. E.g., when a web project uses IIS or IIS Express for debugging, the profile icon could highlight that fact.

If the command name is not specified in the metadata than the page is ignored--without the command name we can't associate it with any actual launch profiles, so there is no point in surfacing it to the user as an option. If the other metadata items are missing we substitute in reasonable defaults.

For testing purposes, this commit also updates our two "commandNameBasedDebugger" pages to specify an ImageMoniker (though it is the same as the default).