Atomically read and update profiles and global settings
The `LaunchSettingsProvider` allows multiple concurrent changes to a project's launch settings by forcing them into a serialized order (via the `SequentialTaskExecutor`). However, two different changes to the same launch profile or global setting could still stomp over each other as there is no way to read and update these settings as a single operation.

Say operation A and operation B want to update different parts of the same launch profile. They each grab the current state, P0. They then each make their own changes to produce PA and PB, respectively. Then they both call `AddOrUpdateProfileAsync` to apply their changes. This method replaces the entire profile at once, so depending on the order in which the calls are processed we'll either end up with PA (in which case B's changes are lost) or PB (A's changes are lost) when what we really want is PAB (with both changes).

Global settings have a similar update mechanism, and a similar problem.

To fix this, this change adds new update methods that take callbacks to make the actual changes. When the update is processed, the `LaunchSettingsProvider` invokes the callback, providing the current state of the profile or global settings. The callback makes its changes, and returns the resulting object which is applied to the settings immediately. Any caller making use of these methods is guaranteed to see the current state of the profile/global setting when they make their changes.

Eventually these methods will be used by the new Launch Settings UI back end, but we should also consider adopting it in other places as well.