Reorder properties on Application page
Historically in legacy projects, the Assembly Name and Default Namespace properties were manually kept in sync with the project by the user.

With SDK-style projects, it's more common to go with the default behaviour of picking up the MSBuildProjectName property, which is computed from the project's file name.

Given these properties are far less likely to be user edited, we move them further down the UI, allowing more commonly modified properties to rise to the top.