Allow updating a profile to fail
Currently, if the profile does not exist at the time the action passed to `UpdateProfileAsync` is to be run it is skipped with no indication to the calling code of what happened. This change renames the method to `TryUpdateProfileAsync` and returns `true` to indicate that the profile was found and the action run, or `false` to indicate the profile was not found.