Add an actual extension value provider
Add an `ILaunchProfileExtensionValueProvider` that handles the extra settings we care about for the "Project" and "Executable" debug commands, along with unit tests.

Note this is _very_ similar to the `ActiveLaunchProfileValueProvider`, except that type can only read and write these properties to the active profile, while the new type can handle any profile given to it. `ActiveLaunchProfileValueProvider` will eventually be deleted.