Fix a potential hang in LaunchSettingsProvider
Quite a few calls into `LaunchSettingsProvider.WaitForFirstSnapshot` pass `Timeout.Infinite`. If such a wait is initiated, and the `LaunchSettingsProvider` is disposed, the wait operation would never complete.

This change ensures the `TaskCompletionSource` via which the wait is implemented is cancelled if it hadn't otherwise completed during disposal of the `LaunchSettingsProvider`, ensuring such hangs can not occur.