Version LaunchSettings
This change introduces a new interface, `IVersionedLaunchSettings`, that returns a version number. The `LaunchSettings` type is updated to implement the new interface, and `LaunchSettingsProvider` now supplies and increments the version every time it produces a new instance of `LaunchSettings`.

Eventually this version will be passed to the Project Query engine so that it knows when it needs to re-run a query to pick up changes.