Support looking up launch profiles by ID
The `LaunchProfileFromProjectDataProducer` can take an `UnconfiguredProject` and produce a set of Project Query API entities representing the launch profiles in that project. However, we currently lack a way to map from an entity ID representing a launch profile to the actual entity. This is necessary in order to support actions on launch profiles via the Project Query API, namely, setting the value of a property in a launch profile. The query engine first tries to look up the launch profile based on its ID, and then applies the action.

The code to create and populate the actual `LaunchProfileValue` is needed in both scenarios (project -> launch profiles, and entity ID -> launch profile) and has been extracted into a new `LaunchProfileDataProducer` type.