Fix dependency tree node customization via extensions
Fixes #6604

Previously `IProjectTreePropertiesProvider` implementations would only be invoked on _updates_ to tree nodes. With this change, they are invoked on creation and updates.

This change also avoids allocating `ProjectTreeCustomizablePropertyContext` and `ReferencesProjectTreeCustomizablePropertyValues` instances when no relevant `IProjectTreePropertiesProvider` imports exist, which will improve performance in the common case.

Adds a documentation file for the dependencies node and adds an example showing how to customize nodes.