Fix duplication of group nodes when filtering captions
#7078 fixed a bug that meant exports of `IProjectTreePropertiesProvider` were not being reliably called when nodes were added to the dependencies tree. This interface is not being used in WAP projects to rename the "Projects" node as "Applications".

When the dependencies tree updates itself, it uses the caption to identify the group node to which the update applies. This search was previously using the caption, and if an extension point renamed the node's caption it would not be found, leading to duplicate nodes appearing in the tree.

This fix introduces an internal `IProjectDependenciesSubTreeProvider2` interface via which our rule handlers may specify a `ProjectTreeFlags` value that uniquely identifies the group node. When this implementation is found, it will be used. If it is not found (for example, the WebTools provider for NPM packages) it falls back to using the caption.

Note that `IProjectTreePropertiesProvider` implementations may also modify a node's flags, however it is considerably less reasonable to remove the uniquely identifying flag from a group node than to change its caption, so this pragmatic approach is taken for now. We can replace this in future if needed, as the new interface is internal.