Move warnaserror+ to the end of the command line arguments.
Current implementation gets the command line arguments from from the rules and this changes the order of the arguments which causes Build to behave differently.
When using the "WarningsNotAsErrors" feature from Intellisense some errors and warnings were showed as warnings in the Error List, but as errors in the Output window.

We can get the correct order of command line arguments by subscribing to ProjectBuildSnapshotService, and then parse the output result from the build target, but since the structure implies iterating over all the elements, extract the arguments, and also changes some interfaces we are  not using this approach, unless we want to preserve ALL the arguments in the same order.

A simple change is to change move 'warnaserror+' to the end of the command line arguments to make sure "WarningsNotAsErrors" behave correctly.