Fix "Startup object" UI
The property pages for executables include a "Startup object" property where you can specify which `Main` method or `Form` to run at startup. In both the old and new UI, the drop down displays all applicable methods and forms. The old UI includes a "(Not set)" item representing the empty value; however, the new UI does not. If you set it to one of the other options there is no way to unset it.

The old UI gets the values for the drop down from the `StartupObjectsEnumProvider`, but then explicitly adds in the "(Not set)" value in the UI code. Since the new UI is generated, this option isn't available to us. Instead, we need `StartupObjectsEnumProvider` to include "(Not set)" along with all the other values, but only when the new UI is in use.

To achieve this we make use of the `DynamicEnumProperty.ProviderSettings`--a bag of name/value pairs that is passed along to the enum provider. ApplicationPropertyPage.xaml is updated to pass along an "IncludeEmptyValue" setting, and `StartupObjectsEnumProvider` is updated to check for the setting.