Keep the order of command line arguments for build.
What are the changes:
Get the command line arguments from IProjectBuildSnapshotService instead of IProjectSubscriptionService.

Subscribe to the ProjectBuildSnapshotService directly using the build targets, instead of rules.

What is the bug:
Getting the data from the rules changes the order of the commands which causes the Build to behave differently.

How this change fixes the bug:
Get the original order from ProjectBuildSnapshotService and pass it directly to roslyn.