Refining the CompoundTargetFrameworkValueProvider.
Refining the CompoundTargetFrameworkValueProvider to correctly read the associated properties and saved them as expected.
These changes also consider the case where the value in the TargetFramework property is user-specific, and not one from the list of SupportedTargetFramework.

At one point, I faced the need to correlate the TargetFrameworkMoniker into the TargetFramework alias (i.e. NETCoreApp,Version=v5.0 -> net5.0), as well as for the platform case to ensure that we save the correct alias into the project file. I did this by calling the project's subscription service to get the set of supported values we already get from the SDK, and then extract the key-values I'm interested in (done on the GetTargetFrameworkAlias and GetTargetPlatformAlias methods).