Remove usage of the 'evaluated' visibility condition function
Fixes #7021

Currently we have a visibility condition function `evaluated` which returns a single evaluated value for a property.

However a project may produce different evaluated values per configuration. Currently, if that happens, the `evaluated` function throws an exception.

We also have an `has-evaluated-value` function which returns true if any of the evaluated value matches an argument. This will never throw the above exception.

The original idea was that we would use `evaluated` for properties which self-identify as non-configurable (and therefore would only have a single value).

The problem is that a project can still be manually constructed that produces different values across configurations for any property. If we use `evaluated` then there will always be a "valid" project that produces an exception in the UI.

This change removes usage of the `evaluated` function, and removes it from the documentation. Once this has merged and inserted, we can remove the definition of this function from CPS.