Create a name if one is not provided
If the `AddLaunchProfile` step does not provide a new profile name we will now generate a unique one. We start with "Profile 1" and then increment the number until we find a name that is not yet used by a launch profile.

The expectation is that the Launch Profiles UI will just create new profiles with default names, and then immediately offer to rename the new profile.