Don't assume en-US culture for customer-facing URLs
When a customer clicks a link to documentation, they expect to see it in their own language (where applicable).

This commit removes the assumption on en-US culture from the URLs we allow customers to navigate to.

More information at https://review.docs.microsoft.com/help/contribute/links-how-to?branch=master#locale-codes-in-external-links