Support filtering up-to-date check items
It may be desirable for a component within Visual Studio to schedule a build for which only a subset of the up-to-date check inputs and outputs should be considered.

To achieve this, we support specifying `Kind` metadata on inputs and items, and check the `FastUpToDateCheckIgnoresKinds` global property for values that indicate which inputs/outputs should be ignored for the purposes of the fast up-to-date check.